# Precompute the indexable range at the sketch level