The indexable range was only precomputed when using the logarithmic mapping.

# Benchmarks

Those include generating the values:

![image](https://user-images.githubusercontent.com/10637860/129177506-0d46628c-b579-4513-bd80-164b6d2dfdef.png)

## Before

```
BenchmarkAdd/logarithmic-12    	22238661	        53.32 ns/op	       0 B/op	       0 allocs/op
BenchmarkAdd/cubic-12          	13639500	        82.57 ns/op	       0 B/op	       0 allocs/op
BenchmarkAdd/linear-12         	13924296	        78.65 ns/op	       0 B/op	       0 allocs/op
```

## After

```
BenchmarkAdd/logarithmic-12         	20049031	        53.22 ns/op	       0 B/op	       0 allocs/op
BenchmarkAdd/cubic-12               	35861640	        32.98 ns/op	       0 B/op	       0 allocs/op
BenchmarkAdd/linear-12              	39687194	        29.46 ns/op	       0 B/op	       0 allocs/op
```